+++ src/google/protobuf/wire_format.h
@@ -268,10 +268,10 @@ class PROTOBUF_EXPORT WireFormat {
   // Verifies that a string field is valid UTF8, logging an error if not.
   // This function will not be called by newly generated protobuf code
   // but remains present to support existing code.
-  static void VerifyUTF8String(const char* data, int size, Operation op);
+  static void VerifyUTF8String(const char* data, size_t size, Operation op);
   // The NamedField variant takes a field name in order to produce an
   // informative error message if verification fails.
-  static void VerifyUTF8StringNamedField(const char* data, int size,
+  static void VerifyUTF8StringNamedField(const char* data, size_t size,
                                          Operation op, const char* field_name);
 
  private:
@@ -337,7 +337,7 @@ inline size_t WireFormat::TagSize(int field_number,
       static_cast<WireFormatLite::FieldType>(implicit_cast<int>(type)));
 }
 
-inline void WireFormat::VerifyUTF8String(const char* data, int size,
+inline void WireFormat::VerifyUTF8String(const char* data, size_t size,
                                          WireFormat::Operation op) {
 #ifdef GOOGLE_PROTOBUF_UTF8_VALIDATION_ENABLED
   WireFormatLite::VerifyUtf8String(
@@ -350,7 +350,7 @@ inline void WireFormat::VerifyUTF8String(const char* data, int size,
 #endif
 }
 
-inline void WireFormat::VerifyUTF8StringNamedField(const char* data, int size,
+inline void WireFormat::VerifyUTF8StringNamedField(const char* data, size_t size,
                                                    WireFormat::Operation op,
                                                    const char* field_name) {
 #ifdef GOOGLE_PROTOBUF_UTF8_VALIDATION_ENABLED
